+++ MK4duo/src/language/language_bg.h
@@ -3,7 +3,7 @@
  *
  * Based on Marlin, Sprinter and grbl
  * Copyright (c) 2011 Camiel Gubbels / Erik van der Zalm
- * Copyright (c) 2019 Alberto Cotronei @MagoKimbra
+ * Copyright (c) 2020 Alberto Cotronei @MagoKimbra
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -111,7 +111,7 @@ namespace language_bg {
   FSTRINGVALUE(MSG_LOAD_EEPROM                      , _UxGT("Зареди от EPROM"));
   FSTRINGVALUE(MSG_RESTORE_FAILSAFE                 , _UxGT("Фабрични настройки"));
   FSTRINGVALUE(MSG_REFRESH                          , LCD_STR_REFRESH _UxGT("Обнови"));
-  FSTRINGVALUE(MSG_WATCH                            , _UxGT("Преглед"));
+  FSTRINGVALUE(MSG_INFO_SCREEN                      , _UxGT("Преглед"));
   FSTRINGVALUE(MSG_PREPARE                          , _UxGT("Действия"));
   FSTRINGVALUE(MSG_TUNE                             , _UxGT("Настройка"));
   FSTRINGVALUE(MSG_PAUSE_PRINT                      , _UxGT("Пауза"));