+++ MK4duo/src/language/language_ca.h
@@ -3,7 +3,7 @@
  *
  * Based on Marlin, Sprinter and grbl
  * Copyright (c) 2011 Camiel Gubbels / Erik van der Zalm
- * Copyright (c) 2019 Alberto Cotronei @MagoKimbra
+ * Copyright (c) 2020 Alberto Cotronei @MagoKimbra
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -122,7 +122,7 @@ namespace language_ca {
   FSTRINGVALUE(MSG_LOAD_EEPROM                      , _UxGT("Carrega memoria"));
   FSTRINGVALUE(MSG_RESTORE_FAILSAFE                 , _UxGT("Restaura valors"));
   FSTRINGVALUE(MSG_REFRESH                          , LCD_STR_REFRESH  _UxGT("Actualitza"));
-  FSTRINGVALUE(MSG_WATCH                            , _UxGT("Pantalla Info."));
+  FSTRINGVALUE(MSG_INFO_SCREEN                      , _UxGT("Pantalla Info."));
   FSTRINGVALUE(MSG_PREPARE                          , _UxGT("Prepara"));
   FSTRINGVALUE(MSG_TUNE                             , _UxGT("Ajusta"));
   FSTRINGVALUE(MSG_PAUSE_PRINT                      , _UxGT("Pausa impressio"));