+++ MK4duo/src/language/language_fr.h
@@ -3,7 +3,7 @@
  *
  * Based on Marlin, Sprinter and grbl
  * Copyright (c) 2011 Camiel Gubbels / Erik van der Zalm
- * Copyright (c) 2019 Alberto Cotronei @MagoKimbra
+ * Copyright (c) 2020 Alberto Cotronei @MagoKimbra
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -285,7 +285,7 @@ namespace language_fr {
   FSTRINGVALUE(MSG_MEDIA_UPDATE                     , _UxGT("MaJ Firmware SD"));
   FSTRINGVALUE(MSG_RESET_PRINTER                    , _UxGT("RaZ imprimante"));
   FSTRINGVALUE(MSG_REFRESH                          , LCD_STR_REFRESH  _UxGT("Actualiser"));
-  FSTRINGVALUE(MSG_WATCH                            , _UxGT("Surveiller"));
+  FSTRINGVALUE(MSG_INFO_SCREEN                      , _UxGT("Surveiller"));
   FSTRINGVALUE(MSG_PREPARE                          , _UxGT("Préparer"));
   FSTRINGVALUE(MSG_TUNE                             , _UxGT("Régler"));
   FSTRINGVALUE(MSG_START_PRINT                      , _UxGT("Démarrer impression"));