+++ MK4duo/src/platform/HAL_DUE/fastio.h
@@ -3,7 +3,7 @@
  *
  * Based on Marlin, Sprinter and grbl
  * Copyright (c) 2011 Camiel Gubbels / Erik van der Zalm
- * Copyright (c) 2019 Alberto Cotronei @MagoKimbra
+ * Copyright (c) 2020 Alberto Cotronei @MagoKimbra
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -28,7 +28,7 @@
  * Contributors:
  *  Copyright (c) 2014 Bob Cousins bobcousins42@googlemail.com
  *  Copyright (c) 2015-2016 Nico Tonnhofer wurstnase.reprap@gmail.com
- *  Copyright (c) 2019 Alberto Cotronei @MagoKimbra
+ *  Copyright (c) 2020 Alberto Cotronei @MagoKimbra
  */
 
 // **************************************************************************